[Improve] maven-wrapper downloader improvement
