--- .mvn/wrapper/MavenWrapperChecker.java
@@ -1,39 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import java.util.jar.JarFile;
-import java.util.zip.ZipException;
-public class MavenWrapperChecker {
-
-    private static final boolean VERBOSE = Boolean.parseBoolean(System.getenv("MVNW_VERBOSE"));
-
-    public static void main(String[] args) {
-        String wrapperJar = args[0];
-        log("maven-wrapper file checking: " + wrapperJar);
-        try(JarFile ignored = new JarFile(wrapperJar, true)) {
-            System.exit(0);
-        } catch (Exception e) {
-            System.exit(1);
-        }
-    }
-    private static void log(String msg) {
-        if (VERBOSE) {
-            System.out.println(msg);
-        }
-    }
-
-}
